+++ esbuild.config.mjs
@@ -115,6 +115,7 @@ async function buildBackground() {
     readFile("modules/sync-action-handler.js"),
     readFile("modules/backup-action-handler.js"),
     readFile("modules/organization-action-handler.js"),
+    readFile("modules/settings-action-handler.js"),
     readFile("modules/message-router.js"),
     readFile("modules/gm-audio-handler.js"),
     readFile("modules/gm-menu-handler.js"),